+++ lib/shameless/configuration.rb
@@ -1,6 +1,6 @@
 module Shameless
   class Configuration
-    attr_accessor :partition_urls, :shards_count
+    attr_accessor :partition_urls, :shards_count, :connection_options
 
     def shards_per_partition_count
       shards_count / partitions_count